Adapted classes to handle authorization

+++ Backend/src/main/java/com/Pubrunda/entities/post/PostService.java
@@ -1,6 +1,9 @@
 package com.Pubrunda.entities.post;
 
+import com.Pubrunda.AuthorizationManager;
 import com.Pubrunda.entities.post.dto.request.PostQueryParams;
+import com.Pubrunda.entities.user.User;
+import com.Pubrunda.exception.AuthorizationException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -22,11 +25,22 @@ public class PostService {
         return postRepository.findAll(postSpecifications);
     }
 
-    public Post createPost(Post newPost) {
+    public Post createPost(User authenticatedUser, Post newPost) {
+        User existingUser = newPost.getAuthor();
+
+        if (!AuthorizationManager.hasAuthorityOfUser(authenticatedUser, existingUser)) {
+            throw new AuthorizationException("You are not allowed to create this post");
+        }
         return postRepository.save(newPost);
     }
 
-    public void deletePost(long postId) {
+    public void deletePost(User authenticatedUser, long postId) {
+        User existingUser = postRepository.findById(postId).orElseThrow().getAuthor();
+
+        if (!AuthorizationManager.hasAuthorityOfUser(authenticatedUser, existingUser)) {
+            throw new AuthorizationException("You are not allowed to delete this post") {
+            };
+        }
         postRepository.deleteById(postId);
     }
